Additional Model Unloading Options
The more I thought about unloading models after they are used, the more I liked it. 
Anyways, my reasoning is that many systems might need the VRAM that the model uses up for the generation of their image. So it would be better to unload the WD model when it is not in use. CLIP model needs to be unloaded from the CLIP interrogator tab.